Refactor image stuff
