+++ tests/content/images/template arg options.html
@@ -1,0 +1,8 @@
+Title: Template argument options (HTML version)
+Date: 2019-04-08 17:07:20-07:00
+Entry-ID: 313
+UUID: 1ba0e0c7-0e54-5988-a80e-2adf73314c95
+
+<p>
+<a href="template-crops.png"><img src="template-crops.png{crop='143,134,350,320',index_crop='593,377,346,300'}"></a>
+</p>
